Changes made as requested
