+++ TeamCode/src/main/java/paths/movements/Path.java
@@ -1,5 +1,7 @@
 package paths.movements;
 
+import org.firstinspires.ftc.teamcode.apexpathing.AutoTest;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -19,14 +21,13 @@ import geometry.Pose;
  * @author DrPixelCat
  * @author Sohum Arora 22985 Paraducks
  */
-public class Path implements FollowerMovement {
+public class Path extends FollowerMovement {
     private final List<String> buildWarnings = new ArrayList<>();
     private final ArrayList<Callback> callbacks = new ArrayList<>();
 
     private PathSegment parametricPath;
     private HeadingInterpolator interpolator;
     private Pose endPose;
-    private boolean started = false, ended = false;
 
     /**
      * Attaches an executable mechanical/software action to this path.
@@ -51,25 +52,6 @@ public class Path implements FollowerMovement {
      * * @return The geometric terminus of the route.
      */
     public Pose getEndPose() { return endPose; }
-
-    @Override
-    public boolean hasStarted() {
-        return started;
-    }
-
-    @Override
-    public boolean hasEnded() {
-        return ended;
-    }
-    @Override
-    public void setStarted(boolean started) {
-        this.started = started;
-    }
-
-    @Override
-    public void setEnded(boolean ended) {
-        this.ended = ended;
-    }
 
     /**
      * Injects the calculated geometric curve (e.g., a B-Spline) that defines
